# doc/api.yaml
openapi: 3.0.3
info:
  title: Simplified login API specification
  description: |-
    This OpenAPI document describes the simplified login API.
    Copy and paste the API from the `paths` key to your OpenAPI document.
  version: "1"
...
paths:
...
  /users/{userId}/photo:
    post:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: userId


      tags: ["Photos"]
      summary: Uploads a photo by the user
      description: |-
        Uploads a photo by the logged user and it appends to the beginning of the list. 

      operationId: uploadPhoto
      requestBody:
        description: User photo upload
        content:
          image/png:
            schema:
              type: string
              format: binary
              description: PNG image file
              minLength: 3
              maxLength: 144


        required: true


      responses:
        '201':
          description: Requested image
          content:
            application/json:
              schema:
                description: Requested image schema
                type: object
                properties:

                  imagePath:
                    description: path to the image
                    type: string
                    minLength: 3
                    maxLength: 100

                  photoID:
                    type: integer
                    description: ID of the photo
                    example: 9123

                  likes:
                    type: array
                    description: An array with the likes of each photo
                    minItems: 0
                    maxItems: 100
                    items:
                      type: integer
                      description: It stores the IDs of the users that liked the photo
                      example: 7

                  comments:
                    type: array
                    description: An array with the comments of each photo
                    minItems: 0
                    maxItems: 100


                    items:
                      type: object
                      properties:
                        user:
                          type: integer
                          description: User ID
                          example: 7


                        comment:
                          type: string
                          description: Comment
                          minLength: 3
                          maxLength: 16
                          pattern: ^[A-Z][0-9]$


                  date:
                    type: string
                    description: Date in which the photo was uploaded
                    minLength: 3
                    maxLength: 16
                    format: date-time

        '400':
          description: Bad Request. Error parsing data or information not valid.

        '401':
          description: Unauthorized. Missing or invalid Authentication token.



        '405':
          description: Method not allowed.
          content:
            text/plain:
              schema:
                type: string
                example: "Method Not Allowed"
                description: Error description
  /users/{userId}/uploadLogo:
    post:
      parameters:
        - name: userId
          schema:
            type: integer
            description: User identifier

          in: path
          required: true
          description: userId


      tags: [ "Photos" ]
      summary: Uploads the profile picture of the user
      description: |-
        Changes the profile picture of the user

      operationId: uploadLogo
      requestBody:
        description: User logo upload
        content:
          image/png:
            schema:
              type: string
              format: binary
              description: PNG image file
              minLength: 3
              maxLength: 160


        required: true


      responses:
        '201':
          description: Logo
          content:
            application/json:
              schema:
                description: Logo schema
                type: object
                properties:

                  imagePath:
                    description: path to the image
                    type: string
                    minLength: 3
                    maxLength: 100

                  photoID:
                    type: integer
                    description: ID of the photo
                    example: 9123

                  likes:
                    type: array
                    description: An array with the likes of the profile picture
                    minItems: 0
                    maxItems: 100
                    items:
                      type: string
                      minLength: 3
                      maxLength: 16

                  comments:
                    type: array
                    description: An array with the comments of the profile picture
                    minItems: 0
                    maxItems: 100


                    items:
                      type: object
                      properties:
                        user:
                          type: integer
                          description: User ID



                        comment:
                          type: string
                          description: Comment
                          minLength: 3
                          maxLength: 16
                          pattern: ^[A-Z][0-9]$


                  date:
                    type: string
                    description: Date in which the photo was uploaded
                    minLength: 3
                    maxLength: 16
                    format: date-time

        '400':
          description: Bad Request. Error parsing data or information not valid.

        '401':
          description: Unauthorized. Missing or invalid Authentication token.



        '405':
          description: Method not allowed.
          content:
            text/plain:
              schema:
                type: string
                example: "Method Not Allowed"
                description: Error description
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []